refactor: 학기 패키지를 domain에서 support로 이동
